itR-14: refactor: move the list to a separate component
